feat(LIVE-9328): split migration
