+++ apps/ledger-live-desktop/src/renderer/screens/exchange/Swap2/Form/SwapWebView.tsx
@@ -48,7 +48,12 @@ export type SwapWebProps = {
   swapState?: Partial<SwapProps>;
 };
 
-export const SWAP_WEB_MANIFEST_ID = "swap-live-app-demo-1";
+export const SwapWebManifestIDs = {
+  Demo0: "swap-live-app-demo-0",
+  Demo1: "swap-live-app-demo-1",
+}
+
+export const SWAP_WEB_MANIFEST_ID = SwapWebManifestIDs.Demo1;
 
 const SwapWebAppWrapper = styled.div<{ isDevelopment: boolean }>(
   ({ isDevelopment }) => `
